refactor: 파일 구조도 정리
