+++ frontend/src/assets/icons/mini_logo.png
