+++ frontend/src/assets/index.ts
@@ -1,6 +1,6 @@
-export { default as logo } from "./icon/logo.svg";
-export { default as mini_logo } from "./icon/mini_logo.png";
-export { default as logout } from "./icon/logout.svg";
+export { default as logo } from "./icons/logo.svg";
+export { default as mini_logo } from "./icons/mini_logo.png";
+export { default as logout } from "./icons/logout.svg";
 
 // 물품 이미지
 export { default as product1 } from "./images/carrier.jpg";
