--- frontend/src/types/example.ts
@@ -1,6 +1,0 @@
-// 예시 타입
-export interface User {
-  id: number;
-  name: string;
-  email: string;
-}
