+++ frontend/src/types/user.ts
@@ -1,0 +1,10 @@
+export interface User {
+  member_id: string;
+  member_name: string;
+  member_pw: string;
+  member_email: string;
+  member_phone: string;
+  member_birth: string;
+  address: string;
+  auth_name: string;
+}
