+++ frontend/src/utils/delay.ts
@@ -1,0 +1,2 @@
+export const delay = (ms: number) =>
+  new Promise(resolve => setTimeout(resolve, ms));
