+++ frontend/src/utils/productImg.ts
@@ -1,0 +1,9 @@
+import { pilow, travelkit, snorkel, umbrella, carrier } from "../assets";
+
+export const productImages = [
+  { id: "carrier", src: carrier },
+  { id: "pilow", src: pilow },
+  { id: "travelkit", src: travelkit },
+  { id: "snorkel", src: snorkel },
+  { id: "umbrella", src: umbrella },
+];
